+++ src/shapeLessRecipe.cpp
@@ -33,7 +33,7 @@ bool shapelessRecipe::verifyPattern(rectangularSlotContainer* container)
 			}
 		}
 	}
-	for (int i = 0; i < requiredItems->size; i++)
+	for (size_t i = 0; i < requiredItems->size; i++)
 	{
 		if (!(found[i]))
 		{
